+++ static/lib/nodep-date-input-polyfill.dist.js
@@ -1,1 +1,0 @@
-../../node_modules/nodep-date-input-polyfill/nodep-date-input-polyfill.dist.js